+++ scripts/ipadapter.py
@@ -33,7 +33,7 @@ class Script(scripts.Script):
                 init_images.append(image)
             except Exception as e:
                 shared.log.warning(f'IP adapter failed to load image: {e}')
-        return init_images
+        return gr.update(value=init_images, visible=len(init_images) > 0)
 
     def display_units(self, num_units):
         num_units = num_units or 1
@@ -47,7 +47,9 @@ class Script(scripts.Script):
             starts = []
             ends = []
             files = []
-            galleries = []
+            masks = []
+            image_galleries = []
+            mask_galleries = []
             with gr.Row():
                 num_adapters = gr.Slider(label="Active IP adapters", minimum=1, maximum=MAX_ADAPTERS, step=1, value=1, scale=1)
             for i in range(MAX_ADAPTERS):
@@ -61,11 +63,16 @@ class Script(scripts.Script):
                     with gr.Row():
                         files.append(gr.File(label='Input images', file_count='multiple', file_types=['image'], type='file', interactive=True, height=100))
                     with gr.Row():
-                        galleries.append(gr.Gallery(show_label=False, value=[]))
-                    files[i].change(fn=self.load_images, inputs=[files[i]], outputs=[galleries[i]])
+                        image_galleries.append(gr.Gallery(show_label=False, value=[], visible=False, container=False, rows=1))
+                    with gr.Row():
+                        masks.append(gr.File(label='Input masks', file_count='multiple', file_types=['image'], type='file', interactive=True, height=100))
+                    with gr.Row():
+                        mask_galleries.append(gr.Gallery(show_label=False, value=[], visible=False))
+                    files[i].change(fn=self.load_images, inputs=[files[i]], outputs=[image_galleries[i]])
+                    masks[i].change(fn=self.load_images, inputs=[masks[i]], outputs=[mask_galleries[i]])
                 units.append(unit)
             num_adapters.change(fn=self.display_units, inputs=[num_adapters], outputs=units)
-        return [num_adapters] + adapters + scales + files + starts + ends
+        return [num_adapters] + adapters + scales + files + starts + ends + masks
 
     def process(self, p: processing.StableDiffusionProcessing, *args): # pylint: disable=arguments-differ
         if shared.backend != shared.Backend.DIFFUSERS:
@@ -84,4 +91,6 @@ class Script(scripts.Script):
             p.ip_adapter_starts = args[MAX_ADAPTERS*3:MAX_ADAPTERS*4][:units]
         if getattr(p, 'ip_adapter_ends', [1.0]) == [1.0]:
             p.ip_adapter_ends = args[MAX_ADAPTERS*4:MAX_ADAPTERS*5][:units]
+        if getattr(p, 'ip_adapter_masks', []) == []:
+            p.ip_adapter_masks = args[MAX_ADAPTERS*5:MAX_ADAPTERS*6][:units]
         # ipadapter.apply(shared.sd_model, p, p.ip_adapter_names, p.ip_adapter_scales, p.ip_adapter_starts, p.ip_adapter_ends, p.ip_adapter_images) # called directly from processing.process_images_inner